+++ gui.cpp
@@ -167,7 +167,7 @@ void scroll_callback(GLFWwindow* window, double xoffset, double yoffset) {
     (void) xoffset;
     GlobalState* state = static_cast<GlobalState*>(glfwGetWindowUserPointer(window));
 
-    state->view_zoom *= pow(1.2, yoffset);
+    state->view_distance *= pow(1.2, -yoffset);
 }
 
 int main() {