+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed obsolete console message when changing focus or timewarp
 - Added numeric simulation of a simple space ship displayed as a 2km sphere
 - Timewarp automatically plateaus when CPU is saturated
+- Changed HUD to display distance from focus and altitude instead of zoom
 
 # v0.2.0 - 2019-08-09
 - Added position markers for celestial bodies